backend/HackaTUM22/requirements.txt
django >= 4.1.3
django_rest_framework >= 0.1.0
requests >= 2.28.1